Use flat source map utils
